Visual Studio 2015 Build Support (#5)
[ADD] Windows Visual Studio 2015 build.

+++ libobjgen/src/Generator.cpp
@@ -97,7 +97,7 @@ std::string Generator::Escape(const std::string& str)
 
     for(auto c : str)
     {
-        if(::isprint(c))
+        if(::isprint(static_cast<uint8_t>(c)))
         {
             switch(c)
             {
